refactor(sql): extract pure INSERT parser helpers
Purpose and Context

The dev-env SQL validator added a small INSERT/REPLACE parser
inside sql.ts. The pure helpers are large enough to warrant a
dedicated module and dedicated unit coverage. Keeping them inline
mixes a stateful check-runner with a self-contained parser, which
makes the parser harder to understand and harder to test.

Key Changes

- Move the pure (state-free) helpers, constants, and interfaces
  to src/lib/validations/sql-insert-parser.ts.
- Replace inline definitions in src/lib/validations/sql.ts with a
  named import from the new module.
- Leave the six module-level currentInsertStatement* variables and
  the five stateful helpers (collectInsertColumnList,
  isSqlCommentOnlyLine, collectOptionUrlMatchesFromRows,
  collectOptionsInsertRows, collectOptionsInsertMatches) in sql.ts
  so this commit is purely a code move.

Impact and Considerations

No behavior change. The extracted helpers have byte-identical
bodies modulo prettier's printWidth: 100 wrap on three signature
lines that crossed the limit once `export ` was prepended. The
existing integration suite at __tests__/lib/validations/sql.js
remains the authoritative behavior-preservation evidence and
passes unchanged (26/26).

Testing and Validation

- npx jest --runTestsByPath __tests__/lib/validations/sql.js --runInBand
- npx eslint src/lib/validations/sql.ts src/lib/validations/sql-insert-parser.ts
- npm run check-types
